Relocate verifier repair decision + policy helpers to verifier_orchestration
Phase 2 batch 6 for Issue #682. Moves 3 verifier-repair policy
helpers (~90 LOC of bodies) out of turn.rs into
verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  - verifier_repair_decision (17 LOC, #[cfg(test)] — thin wrapper
    around `super::repair_job::verifier_repair_decision` retained
    for the legacy test surface).
  - verifier_repair_policy_for_decision (35 LOC, #[cfg(test)] —
    decision -> EffectiveToolPolicy mapping consumed by test fixtures
    pinning the focused-edit policy shape).
  - verifier_repair_policy_for_target_hint (38 LOC, production —
    consumed by the actor loop dispatcher to assemble the verifier-
    repair tool policy from a recovered target hint).

All 3 promoted to pub(super) (file-private before). The 2 cfg(test)
fns retain their #[cfg(test)] attribute since they were test-only
in turn.rs.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::repair_job::{safe_relative_path_string,
      VerifierRepairDecision (#[cfg(test)])}
    - super::tool_history::focused_edit_target_already_read
    - super::tool_policy::{EffectiveToolPolicy,
      EffectiveToolPolicyReason}

  * turn.rs:
    - Extended use super::verifier_orchestration::{...} with
      verifier_repair_policy_for_target_hint (production).
    - Dropped verifier_repair_decision +
      verifier_repair_policy_for_decision +
      verifier_repair_policy_for_target_hint from the
      mod progress_tests use super::{...} and added them to the
      existing use super::super::verifier_orchestration::{...} block.
    - Rewrote 1 super::verifier_repair_decision(...) call site at
      L28427 to super::super::verifier_orchestration::
      verifier_repair_decision(...).
    - Cleaned 2 orphaned #[cfg(test)] attributes left over the
      const PYTHON_REQUEST_PATTERNS line (would otherwise cfg-gate
      the next production fn out of release builds).

Metrics: turn.rs 34,086 -> 33,998 (-88); verifier_orchestration.rs
973 -> 1,072 (+99). Cumulative Issue #682 progress:
turn.rs 34,958 -> 33,998 (-960, ~32% of Phase 2 target).
Acceptance gap: 33,998 -> 32,000 = ~1,998 LOC remaining.
